test(animation): migrate
