+++ core/src/utils/animation/test/multiple/animation.e2e.ts
@@ -1,24 +1,21 @@
 import { expect } from '@playwright/test';
-import { test } from '@utils/test/playwright';
+import { configs, test } from '@utils/test/playwright';
 import type { E2EPage } from '@utils/test/playwright';
 
-test.describe('animation: multiple', async () => {
-  test.beforeEach(({ skip }) => {
-    skip.rtl();
-    skip.mode('ios');
-  });
+configs({ modes: ['md'], directions: ['ltr'] }).forEach(({ title, config }) => {
+  test.describe(title('animation: multiple'), async () => {
+    test(`should resolve grouped animations using web animations`, async ({ page }) => {
+      await page.goto('/src/utils/animation/test/multiple', config);
+      await testMultiple(page);
+    });
 
-  test(`should resolve grouped animations using web animations`, async ({ page }) => {
-    await page.goto('/src/utils/animation/test/multiple');
-    await testMultiple(page);
-  });
-
-  /**
-   * CSS animations will occasionally resolve out of order, so we skip for now
-   */
-  test.skip(`should resolve grouped animations using css animations`, async ({ page }) => {
-    await page.goto('/src/utils/animation/test/multiple?ionic:_forceCSSAnimations=true');
-    await testMultiple(page);
+    /**
+     * CSS animations will occasionally resolve out of order, so we skip for now
+     */
+    test.skip(`should resolve grouped animations using css animations`, async ({ page }) => {
+      await page.goto('/src/utils/animation/test/multiple?ionic:_forceCSSAnimations=true', config);
+      await testMultiple(page);
+    });
   });
 });
